4.17.1.8: move various doc/jove.rc defauls (process-prompt, allow-^S-and-^Q, mode-line, disable-biff, pause-jove binding to ^[s ) to the code, they've been in doc/jove.rc for so long they can now be considered official, it reduces the number of weird differences when one just starts up the binary without SHAREDIR installed or for test.
also leave allow-^S-and-^Q on for vt100.  I refuse to
believe any such real terminals are operational and used for
Jove, but far more terminal emulators (e.g. NetBSD console!)
claim to be vt100.

update ModeFmt and MAX_TYPEOUT to new MAXCOLS to support the
sort of wide-terminal that I suspect most of us now use.  I
think jove can afford the couple of hundred extra bytes
(and when ported back to the PDP-11, MAXCOLS can be set lower,
we can accept that fix when it is tested)